fix(tests): fixed imports and features for API tests
